Move non-task into main lib/ directory
While adding a new rake task it confused me to find a non-task file in
this directory. This moves the file into the main 'lib/' directory,
noting that we already have a precedent for this (whitehall import).
